# openapi.yaml
openapi: 3.1.0
info:
  title: Lunaria API
  version: 1.0.0
  description: |
    Lunaria Discord コミュニティ管理プラットフォームの REST API です。
    すべてのレスポンスは `{ success: true, data: T }` または `{ success: false, error: { code, message } }` の形式で返却されます。
servers:
  - url: http://localhost:4000/api/v1
    description: ローカル開発環境
  - url: https://api.lunaria.example.com/api/v1
    description: 本番環境
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: lunaria_token
...
security:
  - cookieAuth: []
paths:
...
  /guilds/{guildId}/analytics/summary:
    get:
      tags: [analytics]
      summary: アナリティクスサマリー
      parameters:
        - in: path
          name: guildId
          required: true
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: サマリー統計
  /guilds/{guildId}/analytics/daily:
    get:
      tags: [analytics]
      summary: 日別アナリティクス
      parameters:
        - in: path
          name: guildId
          required: true
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 日別データ

  # ---- Audit Logs ----
...
  /guilds/{guildId}/analytics/stream:
    get:
      tags: [analytics]
      summary: Analytics summary の SSE ストリーム
      parameters:
        - in: path
          name: guildId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: text/event-stream
  /guilds/{guildId}/dashboard/widgets:
    get:
      tags: [guilds]
      summary: RBAC 反映済み dashboard widget 一覧
      parameters:
        - in: path
          name: guildId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: widget list
  /guilds/{guildId}/dashboard/layout:
    get:
      tags: [guilds]
      summary: dashboard layout 取得
      parameters:
        - in: path
          name: guildId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: current layout
    put:
      tags: [guilds]
      summary: dashboard layout 更新
      parameters:
        - in: path
          name: guildId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: updated layout
  /guilds/{guildId}/dashboard/layout/reset:
    post:
      tags: [guilds]
      summary: dashboard layout を初期化
      parameters:
        - in: path
          name: guildId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: reset layout